requirements.txt
Babel==1.3
Flask==0.10.1
Flask-Babel==0.9
Flask-Bootstrap==3.3.2.1
Flask-Login==0.2.11
Flask-Mail==0.9.1
Flask-OpenID==1.2.4
Flask-Principal==0.4.0
Flask-SQLAlchemy==2.0
Flask-Security==1.7.4
Flask-Storage==0.1.0
-e git://github.com/FelixLoether/flask-uploads@9e1266fcc09b344caf3b8f4859304813d17448d0#egg=Flask_Uploads-master
Flask-WTF==0.11
Flask-WhooshAlchemy==0.56
Jinja2==2.7.3
MarkupSafe==0.23
ProxyTypes==0.9
SQLAlchemy==0.9.9
Tempita==0.5.2
WTForms==2.0.2
Werkzeug==0.10.1
Whoosh==2.6.0
argparse==1.2.1
blinker==1.3
coverage==3.7.1
decorator==3.4.0
flipflop==1.0
guess-language==0.2
itsdangerous==0.24
passlib==1.6.2
pbr==0.10.8
python-loaders==0.2.3
python-openid==2.2.5
pytz==2014.10
six==1.9.0
speaklater==1.3
sqlalchemy-migrate==0.9.5
sqlparse==0.1.14
wsgiref==0.1.2
